v 1.11 - PapelMoeda
Alterada a função getQuantidade, mudando o tipo primitivo do retorno para float.

+++ Source Code Inspection/src/br/calebe/ticketmachine/core/PapelMoeda.java
@@ -14,7 +14,7 @@ public class PapelMoeda {
         this.quantidade = 0;
     }
 
-    public int getValor() {
+    public float getValor() {
         return valor;
     }
